@jberthold 2024-12-19
This branch seems to fix the issues reported by @PetarMax  for the Lido proof with parallel exploration.
Running standard KEVM and Kontrol performance tests (KEVM with a version using the updated LLVM backend that was rolled back).
Once we are convinced this does not cause issues on larger proofs we can merge this.